+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from "styled-components";
-import Home from "./Home";
+import Home from "./pages/Home";
 import { BrowserRouter as Router, Routes, Route} from "react-router-dom";
 
 function App() {
